+++ R/R62Fun.R
@@ -3,6 +3,7 @@
 #' @description Auto-generates methods from an R6 Class.
 #' @param R6Class R6ClassGenerator to generate public methods from
 #' @param assignEnvir environment in which to assign the S4 generics/methods, default is parent of current environment.
+#' @param detectGeneric logical, if TRUE (default) detects if the method has a S3 generic and defines function accordingly
 #' @details Searches in a given R6 class for all public methods that are not 'initialize' or 'clone'.
 #' For each method if a generic does not already exist, one is created and assigned to the given environment.
 #' Methods are created for every generic, following standard S4 convention.
@@ -17,7 +18,8 @@
 #' printer(pm, "Test String B")
 #'
 #' @export
-R62Fun <- function(R6Class, assignEnvir = parent.env(environment())){
+R62Fun <- function(R6Class, assignEnvir = parent.env(environment()),
+                   detectGeneric = TRUE){
 
   checkmate::assert(inherits(R6Class,"R6ClassGenerator"),
                     .var.name = "R6Class must be an R6ClassGenerator")
@@ -27,7 +29,22 @@ R62Fun <- function(R6Class, assignEnvir = parent.env(environment())){
 
   if(length(methods)>0){
     for(i in 1:length(methods)){
-      methodname = names(methods)[[i]]
+      if(detectGeneric){
+        generic = FALSE
+        x = tryCatch(methods(methodname),warning = function(w) w, error = function(e) e)
+        if(inherits(x, "condition")){
+          if(!grepl("appears not to be S3 generic",x$message) & !inherits(x, "error"))
+            generic = TRUE
+        } else if(length(x)!=0)
+          generic = TRUE
+
+        if(generic)
+          methodname = paste(names(methods)[[i]],R6Class$classname,sep=".")
+        else
+          methodname = names(methods)[[i]]
+      } else
+        methodname = names(methods)[[i]]
+
       value = function(object){}
       formals(value) = c(formals(value), formals(methods[[i]]))
       body(value) = substitute({